requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
faiss-cpu>=1.7.4
transformers>=4.35.0
sentence-transformers>=2.2.2
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.6.2
streamlit>=1.30.0
pypdf>=3.0.0
python-dotenv>=1.0.0
